tests/integration/apps/postgres-app/requirements.txt
starlette
uvicorn
psycopg2-binary
psycopg[binary]
asyncpg
aiopg
